+++ src/hmi/commands_mqtt.cpp
@@ -7,7 +7,7 @@
 #include "commands_engine.h"
 #include "./libs_3rd_party/ArduinoJson-v6.14.1/ArduinoJson-v6.14.1.h"
 #include "./mqtt/mqtt_client.h"
-#include "./mqtt/mqtt_RemoteControl.h"
+#include "./mqtt/mqtt_remoteControl.h"
 #include "./system/system_configuration.h"
 #include "./system/utils.h"
